requirements.txt
Flask
flask_cors
celery
redis
pillow
pytesseract
boto3
flask_sqlalchemy